CEHCIG-116 Adding DatasetReferenceDate and upgrading libraries.
- DatasetReferenceDate added
- Save button detects if save required
- Exit button detects if there are unsaved records
- Delete button asks for confirmation
- upgrade all Bower libraries, especially Underscore and Backbone.
- Dropping use or the Underscore-amd anf Backbone-amd libraries as they
  were not being maintained with the latest releases of those libraries

+++ web/src/scripts/views/EditorView.coffee
@@ -1,10 +1,10 @@
 define [
   'underscore'
-  'jquery'
   'backbone'
   'tpl!templates/Editor.tpl'
   'cs!views/MessageView'
   'cs!views/editor/TitleView'
+  'cs!views/editor/SingleObjectView'
   'cs!views/editor/ResourceTypeView'
   'cs!views/editor/AlternateTitlesView'
   'cs!views/editor/DescriptionView'
@@ -12,10 +12,11 @@ define [
   'cs!views/editor/TopicCategoriesView'
   'cs!views/editor/ContactsView'
   'cs!views/editor/ResourceIdentifiersView'
-], (_, $, Backbone, template, MessageView, TitleView, ResourceTypeView, AlternateTitlesView, DescriptionView, LineageView, TopicCategoriesView, ContactsView, ResourceIdentifiersView) -> Backbone.View.extend
+  'cs!views/editor/DatasetReferenceDateView'
+], (_, Backbone, template, MessageView, TitleView, SingleObjectView, ResourceTypeView, AlternateTitlesView, DescriptionView, LineageView, TopicCategoriesView, ContactsView, ResourceIdentifiersView, DatasetReferenceDateView) -> Backbone.View.extend
 
   events:
-    'click #editorDelete': 'delete'
+    'click #editorDelete': 'attemptDelete'
     'click #editorExit': 'exit'
     'click #editorSave': 'save'
     'click #editorBack': 'back'
@@ -26,47 +27,88 @@ define [
     @listenTo @model, 'error', (model, response) ->
       alert "Problem communicating with server: #{response.status} (#{response.statusText})"
 
+    @listenTo @model, 'change save:required', @toggleSave
+
     @currentStep = 1
     do @render
     @components = [
       new TitleView
         el: @$('#editorTitle')
-        model: @model,
-      new ResourceTypeView
+        model: @model
+
+      new SingleObjectView
         el: @$('#editorResourceType')
-        model: @model,
+        model: @model
+        modelAttribute: 'resourceType'
+        label: 'Resource Type'
+        ObjectInputView: ResourceTypeView,
+        helpText: """
+                  Type of resource.
+                  """
+
+      new SingleObjectView
+        el: @$('#editorDatasetReferenceDate')
+        model: @model
+        modelAttribute: 'datasetReferenceDate'
+        label: 'Dataset Reference Date'
+        ObjectInputView: DatasetReferenceDateView,
+        helpText: """
+                  <p>Creation date, the date the data resource is created.</p>
+                  <p>The publication date is the date when the data resource is being made available or released for use - it is <strong>NOT</strong> the date of creation.</p>
+                  <p>If you include a revision date, it implies that the resource has been changed as a consequence of edits or updates.  For EIDC Hub records it is usual practice for revised resources to have an entirely new record, therefore <em>revision date</em> is rarely necessary.</p>
+                  """
+
       new AlternateTitlesView
         el: @$('#editorAlternateTitles')
-        model: @model,
+        model: @model
+
       new DescriptionView
         el: @$('#editorDescription')
-        model: @model,
+        model: @model
+
       new LineageView
         el: @$('#editorLineage')
-        model: @model,
+        model: @model
+
       new TopicCategoriesView
         el: @$('#editorTopicCategories')
-        model: @model,
+        model: @model
+
       new ContactsView
         el: @$('#editorContacts')
-        model: @model,
+        model: @model
+
       new ResourceIdentifiersView
         el: @$('#editorResourceIdentifiers')
         model: @model
     ]
 
-  delete: ->
-    do @model.destroy
+  toggleSave: ->
+    @$('#editorSave').prop 'disabled', (i, current) -> not current
+
+  attemptDelete: ->
+    if confirm "Delete metadata?"
+      do @model.destroy
 
   save: ->
     @model.save {},
+      success: ->
+        do @toggleSave
       error: (model, response) ->
         console.log "Error saving metadata: #{response.status} (#{response.statusText})"
 
   exit: ->
-    _.invoke @components, 'remove'
-    do @remove
-    do Backbone.history.location.reload
+    reallyExit = =>
+      _.invoke @components, 'remove'
+      do @remove
+      do Backbone.history.location.reload
+
+    if not @$('#editorSave').prop 'disabled'
+      if confirm "Exit without saving changes to metadata?"
+        do reallyExit
+    else
+      do reallyExit
+
 
   back: ->
     @navigate @currentStep - 1
@@ -84,30 +126,31 @@ define [
     @navigate step
 
   navigate: (newStep) ->
-    $nav = $('#editorNav li')
+    $nav = @$('#editorNav li')
     maxStep = $nav.length
     @currentStep = newStep
     @currentStep = 1 if @currentStep < 1
     @currentStep = maxStep if @currentStep > maxStep
 
-    $back = $('#editorBack')
+    $back = @$('#editorBack')
     if @currentStep == 1
       $back.prop 'disabled', true
     else
       $back.prop 'disabled', false
 
-    $next = $('#editorNext')
+    $next = @$('#editorNext')
     if @currentStep == maxStep
       $next.prop 'disabled', true
     else
       $next.prop 'disabled', false
 
     $nav.filter('.active').toggleClass 'active'
-    $($nav[@currentStep - 1]).toggleClass 'active'
+    @$($nav[@currentStep - 1]).toggleClass 'active'
 
-    $step = $('#editor .step')
+    $step = @$('.step')
     $step.filter('.visible').toggleClass 'visible'
-    $($step[@currentStep - 1]).toggleClass 'visible'
+    @$($step[@currentStep - 1]).toggleClass 'visible'
 
   render: ->
     @$el.html template
+    @
